Implement refactored user modal
