+++ frontend/src/app/admin/newUserModal/newUserModal.module.css
@@ -1,0 +1,14 @@
+.subheader {
+  font-size: 20px;
+  font-weight: bold;
+}
+
+.modalInput {
+  width: 100%;
+  margin: 15px 0;
+  padding: 0 10px;
+  border: 1px solid rgba(var(--foreground-rgb), 0.5);
+  background-color: rgb(var(--background-rgb));
+  border-radius: 4px;
+  height: 40px;
+}
