+++ frontend/src/app/admin/page.tsx
@@ -1,11 +1,3 @@
-import Dashboard from "@/components/adminDashboard/adminDashboard";
-import Navbar from "@/components/navbar/navbar";
-
 export default function Admin() {
-  return (
-    <div>
-      <Navbar page="admin" />
-      <Dashboard />
-    </div>
-  );
+  return <div />;
 }
